requirements.txt
argparse
jsonschema
openai
python-dotenv
spacy
titlecase
unidecode
vertexai